requirements.txt
anthropic==0.43.1
llava==1.7.0.dev0
mantis_vl==0.0.5
matplotlib==3.10.0
openai==1.59.8
OpenCC==1.1.9
Pillow==11.1.0
Pinyin2Hanzi==0.1.1
pypinyin==0.53.0
qwen_vl_utils==0.0.8
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.46.3
